Refactor useComponentLinkParams, useProjectComponents, and useCurrentData to avoid react lint errors; ignore more MopedDataGrid errors
