+++ editor/src/views/projects/projectView/ProjectFunding/ProjectFundingTable.jsx
@@ -163,6 +163,7 @@ const ProjectFundingTable = ({
   };
 
   useEffect(() => {
+    // eslint-disable-next-line react-hooks/set-state-in-effect -- rows state is also used for optimistic updates after mutations; refactor in issue #28901
     setRows(tableFundingRows);
   }, [tableFundingRows]);
