+++ editor/src/views/projects/projectView/ProjectMilestones/MilestoneTemplateModal.jsx
@@ -81,6 +81,7 @@ const MilestoneTemplateModal = ({
   );
 
   useEffect(() => {
+    // eslint-disable-next-line react-hooks/set-state-in-effect -- rows state is also used for optimistic updates after mutations; refactor in issue #28901
     setMilestonesToAdd([...filteredMilestonesList]);
   }, [filteredMilestonesList]);
