+++ editor/src/views/projects/projectView/ProjectNotes.jsx
@@ -24,6 +24,7 @@ import NoteInput from "src/views/projects/projectView/ProjectNotes/NoteInput";
 import NoteTypeButton from "src/views/projects/projectView/ProjectNotes/NoteTypeButton";
 import DeleteConfirmationModal from "src/views/projects/projectView/DeleteConfirmationModal";
 import ProjectNote from "src/views/projects/projectView/ProjectNotes/ProjectNote";
+import { useNoteTypeObject } from "src/views/projects/projectView/ProjectNotes/useNoteTypeObject";
 
 import * as yup from "yup";
 import { yupValidator } from "src/utils/validation";
@@ -44,24 +45,6 @@ const validationSchema = yup.object().shape({
   projectStatusUpdate: agolValidation.projectStatusUpdate,
 });
 const validator = (value) => yupValidator(value, validationSchema);
-
-/**
- * Hook to create an object mapping note type slugs to their IDs
- * @param {Array} noteTypes - Array of note types from moped_note_types query
- * @returns
- */
-export const useNoteTypeObject = (noteTypes) =>
-  useMemo(
-    () =>
-      noteTypes.reduce(
-        (obj, item) =>
-          Object.assign(obj, {
-            [item.slug]: item.id,
-          }),
-        {}
-      ),
-    [noteTypes]
-  );
 
 /**
  * Hook to filter notes based on the selected note type and if eCAPRIS status syncing is enabled
